GD-22: Calling Invoke on `SceneRunner` must propagate exceptions (#65)
# Why
When using `Invoke` on the `SceneRunner`, no exceptions are forwarded
because Godot `call` does not transfer the exceptions to C# scripts.

# What
- Refactor the `Invoke` extension do handle it explicitly for
`GDScripts` and `CSharpScripts`.
- Fix handle calling static methods when invoke
- `SceneRunner::Invoke` forward to  `Invoke` extension implementation
- Added new `InvokeAsync` to the `SceneRunner` to support async calls
- Fix invalid Variant unboxing of integer and float
- Improved the `ExceptionAssert` by providing new methods to verify for
exception file line number and file name

+++ test/src/asserts/AssertionsTest.cs
@@ -20,11 +20,11 @@ public class AssertionsTest
     public void AssertThatVariants()
     {
         // Godot number Variants
-        AssertObject(AssertThat(Variant.From((sbyte)-1))).IsInstanceOf<INumberAssert<long>>();
-        AssertObject(AssertThat(Variant.From(11))).IsInstanceOf<INumberAssert<long>>();
-        AssertObject(AssertThat(Variant.From(12L))).IsInstanceOf<INumberAssert<long>>();
-        AssertObject(AssertThat(Variant.From(1.4f))).IsInstanceOf<INumberAssert<double>>();
-        AssertObject(AssertThat(Variant.From(1.5d))).IsInstanceOf<INumberAssert<double>>();
+        AssertObject(AssertThat(Variant.From((sbyte)-1))).IsInstanceOf<INumberAssert<int>>();
+        AssertObject(AssertThat(Variant.From(11))).IsInstanceOf<INumberAssert<int>>();
+        AssertObject(AssertThat(Variant.From(12L))).IsInstanceOf<INumberAssert<int>>();
+        AssertObject(AssertThat(Variant.From(1.4f))).IsInstanceOf<INumberAssert<float>>();
+        AssertObject(AssertThat(Variant.From(1.5d))).IsInstanceOf<INumberAssert<float>>();
     }
 
     [TestCase]
